setting up the "new" game with the new classes

+++ src/entity/Position.java
@@ -2,10 +2,10 @@ package entity;
 
 public class Position {
     
-    private float x;
-    private float y;
+    private int x;
+    private int y;
 
-    Position(float x, float y) {
+    Position(int x, int y) {
         this.x = x;
         this.y = y;
     }
@@ -14,17 +14,17 @@ public class Position {
     
 
 
-    public float getX() {
+    public int getX() {
         return x;
     }
-    public void setX(float x) {
+    public void setX(int x) {
         this.x = x;
     }
 
-    public float getY() {
+    public int getY() {
         return y;
     }
-    public void setY(float y) {
+    public void setY(int y) {
         this.y = y;
     }
